Unfnest/requirements.txt
PySide6>=6.5.0
ezdxf>=1.0.0
requests>=2.31.0
shapely>=2.0.0
numpy>=1.20.0
scipy>=1.7.0
Pillow>=9.0.0
